Rename strategies files to homemade to match docs
